Re-factor Drag and Drop Logic, New File
