+++ client/components/Category.jsx
@@ -9,7 +9,6 @@ export default function Category({
   category,
   categoryId,
   addNewTask,
-
   editTask,
   reRender,
 }) {
@@ -107,7 +106,6 @@ export default function Category({
 
   return (
     <div>
-      {/* UPDATE TITLE HERE */}
       {isEditing ? (
         <input
           type="text"
@@ -150,11 +148,6 @@ export default function Category({
             }}
             className="columnShadow"
           >
-            {' '}
-            <div>
-              {console.log('Logging category.items:', category.items) || null}
-              {/* rest of your code */}
-            </div>
             {category.items.map((task, index) => (
               <Task
                 key={task._id}
